Suspending work on this branch.
Debugging problems with overrides is difficult because SAW's error
messages are not always helpful. It seems more pragmatic at this point
to implement assembly routines with C stubs instead of specifying them
with overrides. The main difficulty is that functions that operate on
variable-size data must be given one override per size.