Giuliano 3 (#5)
Add memory safety proofs for blst_keygen, hash_to_field, sk_to_pk,  group checks, and deserialize.
This excludes assembly routines, which are given a spec and assumed correct for now. One exception is blst_sha256_bcopy: verification fails for an unknown reason using a spec, and thus we give it a C implementation for now.
Note that the build script applies some patches to the C source before verification.