Make proof of `mul_to_fp_thm` work with `z3-4.8.10`.
There seems to be some performance regression for z3 on proof goals
involving the bv2nat operator, which caused the `mul_to_fp_thm`
proof to fail to terminate with `z3-4.8.10`.

The proof has now been refactored to use some lemmas as rewrite
rules, which makes it possible to solve with 4.8.10.