Two updates to the SAW integration:
1. Update the SAW download to use the latest release instead of an
ancient nightly that's no longer available.

2. Skip the downloads of both SAW and Cryptol if IN_SAW_CI is set
to "yes".

These are motivated by the use of this tree in the SAW CI tests, which
has been blowing up, first because we tried to delete the ancient
nightlies, then because the website reorg broke the downloads.

However, at least point (1) seems to be of general value.